feat: 링크 추가 컴포넌트 관련 구조 변경
